Improve backend to work correctly with regtype field and GUI to read any amount of fields

+++ src/libgptbackend/include/registry.h
@@ -38,13 +38,13 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 extern "C" {
 #endif /* __cplusplus */
 
-const char *regtype2str(uint16_t &regtype);
-uint16_t str2regtype(const char *regtype);
+const char *regtype2str(uint32_t &regtype);
+uint32_t str2regtype(const char *regtype);
 
 struct registry_entry {
     char *keyname;
     char *valuename;
-    uint16_t regtype;
+    uint32_t regtype;
     char *data;
 };
